separate models to improve file structure
